+++ src/apps/components/HorizontalSpacer/index.tsx
@@ -1,0 +1,2 @@
+export * from "./HorizontalSpacer";
+export { default } from "./HorizontalSpacer";
