--- src/apps/components/Spacer/index.tsx
@@ -1,2 +1,0 @@
-export * from "./Spacer";
-export { default } from "./Spacer";
