--- src/attributes/components/AttributeSwatchField/messages.ts
@@ -1,12 +1,0 @@
-import { defineMessages } from "react-intl";
-
-export const defined = defineMessages({
-  picker: {
-    defaultMessage: "Picker",
-    description: "swatch attribute color picker label"
-  },
-  image: {
-    defaultMessage: "Image",
-    description: "swatch attribute image label"
-  }
-});
